Fix bug with conditional branching
the register values were not being cast to signed integer types, resulting in incorrect evaluation

+++ sim/sim.c
@@ -174,23 +174,23 @@ int SimulateInstruction(union mips_instruction* inst, struct virtual_mem_region*
 		case OP_BLTZ_BGEZ:
 			switch (inst->itype.rt) {
 				case BRANCH_BLTZ:
-					if (ctx->regs[inst->itype.rs] < 0) {
+					if ((int32_t)ctx->regs[inst->itype.rs] < 0) {
 						ctx->pc += (int16_t)inst->itype.imm << 2;
 					}
 					return 1;
 				case BRANCH_BGEZ:
-					if (ctx->regs[inst->itype.rs] >= 0) {
+					if ((int32_t)ctx->regs[inst->itype.rs] >= 0) {
 						ctx->pc += (int16_t)inst->itype.imm << 2;
 					}
 					return 1;
 				case BRANCH_BLTZAL:
-					if (ctx->regs[inst->itype.rs] < 0) {
+					if ((int32_t)ctx->regs[inst->itype.rs] < 0) {
 						ctx->regs[ra] = ctx->pc + 4;
 						ctx->pc += (int16_t)inst->itype.imm << 2;
 					}
 					return 1;
 				case BRANCH_BGEZAL:
-					if (ctx->regs[inst->itype.rs] >= 0) {
+					if ((int32_t)ctx->regs[inst->itype.rs] >= 0) {
 						ctx->regs[ra] = ctx->pc + 4;
 						ctx->pc += (int16_t)inst->itype.imm << 2;
 					}
